blast/orig_requirements.txt
amqp==1.4.9
anyjson==0.3.3
apns2==0.3.0
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.23
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.1
cryptography==2.1.4
defusedxml==0.5.0
Django==1.9.1
django-allauth==0.29.0
django-appconf==1.0.2
django-contrib-comments==1.8.0
django-crispy-forms==1.7.0
django-filter==0.12.0
django-imagekit==3.3
django-model-utils==3.1.1
django-notifications-hq==1.3
django-pagination-bootstrap==1.2.4
django-pure-pagination==0.3.0
django-push-notifications==1.4.1
django-rest-auth==0.9.3
django-rest-swagger==0.3.8
djangorestframework==3.3.3
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.8.0
djangorestframework-oauth==1.1.0
ecdsa==0.13
first==2.0.1
h2==2.6.2
hpack==3.0.0
http-ece==1.0.5
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
kombu==3.0.35
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
pilkit==1.1.13
Pillow==2.9.0
psycopg2==2.6.1
py-vapid==1.3.0
pycparser==2.18
PyJWT==1.4.0
python3-openid==3.1.0
pytz==2016.6
pywebpush==1.6.0
PyYAML==3.11
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.8.0
simplejson==3.13.2
sinchsms==1.0.3
six==1.10.0
typing==3.5.2.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
